refactor: split eval/run_eval.py scoring into eval/scorers/ subpackage
Closes #259

eval/run_eval.py was 1,563 lines mixing per-case scoring with orchestration.
Per-case scorers (citation grounding, claim alignment, answer format, chunk
metrics, score_case) move to a new eval/scorers/ subpackage, leaving
run_eval.py with config loading, aggregation, and the CLI driver.

eval/scorers/_shared.py holds the helpers used by both scorers and
orchestration (contains_all_terms, hardcase_categories, canonical_query_type,
retry_trigger_reasons, answer_* extraction). Placing them in a leaf module
avoids the circular import that would otherwise arise between run_eval.py
and eval/scorers/__init__.py.

Behavior preserved end-to-end: `make smoke` env-invariant subset of
reports/eval_summary.json is byte-identical between main and this branch.
