+++ eval/run_eval.py
@@ -4,13 +4,11 @@ from collections import Counter, defaultdict
 import datetime
 import hashlib
 import json
-import math
 from pathlib import Path
 import re
 import subprocess
 import sys
 from typing import Any
-import unicodedata
 
 import yaml
 
@@ -30,38 +28,25 @@ from rag_core import (
     run_rag_query,
 )
 from eval.bootstrap import bootstrap_ci
+from eval.scorers import derive_gold_chunk_ids, score_case
+from eval.scorers._shared import (
+    QUERY_TYPE_ALIASES,
+    answer_status,
+    canonical_query_type,
+    hardcase_categories,
+    retry_trigger_reasons,
+)
+from eval.scorers.citation import is_bbox
 from scripts.write_real_eval_baseline import provenance
 
 
 QUERY_TYPES = ("single_doc", "comparison", "follow_up", "abstention")
-QUERY_TYPE_ALIASES = {"multi_doc": "comparison"}
 DEFAULT_ABLATION_RUNS = [
     {
         "name": DEFAULT_CLI_PIPELINE_NAME,
         "pipeline": DEFAULT_CLI_PIPELINE_NAME,
     }
 ]
-
-
-def hardcase_categories(item: dict[str, Any]) -> list[str]:
-    categories = item.get("hardcase_categories") or item.get("hardcase_category") or []
-    if isinstance(categories, str):
-        categories = [categories]
-    if not isinstance(categories, list):
-        return []
-    normalized = []
-    seen: set[str] = set()
-    for category in categories:
-        value = str(category).strip()
-        if value and value not in seen:
-            normalized.append(value)
-            seen.add(value)
-    return normalized
-
-
-def canonical_query_type(query_type: Any) -> str:
-    value = str(query_type or "").strip()
-    return QUERY_TYPE_ALIASES.get(value, value)
 
 
 def _git(*args: str) -> str:
@@ -251,737 +236,6 @@ def load_config(path: Path) -> dict[str, Any]:
             raise ValueError(f"Duplicate ablation run name: {normalized_run['name']}")
         seen_names.add(normalized_run["name"])
     return data
-
-
-def contains_all_terms(text: str, terms: list[str]) -> bool:
-    lowered = text.lower()
-    return all(str(term).lower() in lowered for term in terms)
-
-
-def retry_trigger_reasons(prediction: dict[str, Any]) -> list[str]:
-    diagnostics = prediction.get("diagnostics") or {}
-    reasons: list[str] = []
-    for attempt in diagnostics.get("filter_stage_attempts") or []:
-        if attempt.get("verified"):
-            continue
-        reasons.extend(str(reason) for reason in attempt.get("verification_reasons") or [])
-    return reasons
-
-
-def answer_payload(prediction: dict[str, Any]) -> dict[str, Any]:
-    answer = prediction.get("answer")
-    return answer if isinstance(answer, dict) else {}
-
-
-def answer_claims(prediction: dict[str, Any]) -> list[dict[str, Any]]:
-    claims = answer_payload(prediction).get("claims") or []
-    return [claim for claim in claims if isinstance(claim, dict)]
-
-
-def answer_to_text(prediction: dict[str, Any]) -> str:
-    payload = answer_payload(prediction)
-    if not payload:
-        return str(prediction.get("answer") or "")
-    parts = [
-        str(payload.get("summary") or ""),
-        str(prediction.get("answer_text") or ""),
-    ]
-    for claim in answer_claims(prediction):
-        parts.extend([str(claim.get("claim") or ""), str(claim.get("support") or "")])
-    insufficiency = payload.get("insufficiency")
-    if isinstance(insufficiency, dict):
-        parts.append(str(insufficiency.get("message") or ""))
-    return " ".join(part for part in parts if part)
-
-
-def answer_status(prediction: dict[str, Any]) -> str:
-    payload = answer_payload(prediction)
-    diagnostics = prediction.get("diagnostics") or {}
-    return str(payload.get("status") or diagnostics.get("answer_status") or "")
-
-
-def answer_citations(prediction: dict[str, Any]) -> list[dict[str, Any]]:
-    citations: list[dict[str, Any]] = []
-    for claim in answer_claims(prediction):
-        for citation in claim.get("citations") or []:
-            if isinstance(citation, dict):
-                citations.append(citation)
-    return citations
-
-
-def is_bbox(value: Any) -> bool:
-    if not isinstance(value, list) or len(value) != 4:
-        return False
-    try:
-        [float(part) for part in value]
-    except (TypeError, ValueError):
-        return False
-    return True
-
-
-def bbox_iou(left: list[Any], right: list[Any]) -> float:
-    l0, t0, l1, b1 = [float(part) for part in left]
-    r0, u0, r1, d1 = [float(part) for part in right]
-    inter_w = max(0.0, min(l1, r1) - max(l0, r0))
-    inter_h = max(0.0, min(b1, d1) - max(t0, u0))
-    intersection = inter_w * inter_h
-    left_area = max(0.0, l1 - l0) * max(0.0, b1 - t0)
-    right_area = max(0.0, r1 - r0) * max(0.0, d1 - u0)
-    union = left_area + right_area - intersection
-    return 0.0 if union <= 0 else intersection / union
-
-
-def citation_pages(citation: dict[str, Any]) -> set[int]:
-    pages: set[int] = set()
-    page_span = citation.get("page_span")
-    if (
-        isinstance(page_span, list)
-        and len(page_span) == 2
-        and all(isinstance(page, int) for page in page_span)
-    ):
-        start, end = page_span
-        if start <= end:
-            pages.update(range(start, end + 1))
-    for region in citation.get("regions") or []:
-        if isinstance(region, dict) and isinstance(region.get("page_number"), int):
-            pages.add(int(region["page_number"]))
-    return pages
-
-
-def score_citation_pages(
-    expected_pages: list[dict[str, Any]],
-    citations: list[dict[str, Any]],
-) -> tuple[float | None, list[dict[str, Any]]]:
-    if not expected_pages:
-        return None, []
-
-    matched = 0
-    errors: list[dict[str, Any]] = []
-    for expected in expected_pages:
-        doc_id = str(expected.get("doc_id") or "")
-        pages = {int(page) for page in expected.get("pages") or [] if isinstance(page, int)}
-        same_doc = [citation for citation in citations if str(citation.get("doc_id") or "") == doc_id]
-        page_sets = [citation_pages(citation) for citation in same_doc]
-        page_sets = [page_set for page_set in page_sets if page_set]
-        if any(page_set & pages for page_set in page_sets):
-            matched += 1
-            continue
-        if not page_sets:
-            errors.append(
-                {
-                    "code": "page_missing",
-                    "message": "Expected citation page metadata was unavailable.",
-                    "expected": {"doc_id": doc_id, "pages": sorted(pages)},
-                }
-            )
-        else:
-            errors.append(
-                {
-                    "code": "page_mismatch",
-                    "message": "Citation page metadata did not overlap expected pages.",
-                    "expected": {"doc_id": doc_id, "pages": sorted(pages)},
-                    "actual_pages": sorted({page for page_set in page_sets for page in page_set}),
-                }
-            )
-    return matched / len(expected_pages), errors
-
-
-def citation_regions(citation: dict[str, Any]) -> list[dict[str, Any]]:
-    regions = citation.get("regions") or []
-    return [region for region in regions if isinstance(region, dict)]
-
-
-def score_citation_regions(
-    expected_regions: list[dict[str, Any]],
-    citations: list[dict[str, Any]],
-) -> tuple[float | None, list[dict[str, Any]]]:
-    if not expected_regions:
-        return None, []
-
-    matched = 0
-    errors: list[dict[str, Any]] = []
-    for expected in expected_regions:
-        doc_id = str(expected.get("doc_id") or "")
-        page_number = int(expected["page_number"])
-        expected_bbox = expected.get("bbox")
-        min_iou = float(expected.get("min_iou", 0.5))
-        candidate_regions: list[dict[str, Any]] = []
-        for citation in citations:
-            if str(citation.get("doc_id") or "") != doc_id:
-                continue
-            for region in citation_regions(citation):
-                if region.get("page_number") == page_number and is_bbox(region.get("bbox")):
-                    candidate_regions.append(region)
-        if not candidate_regions:
-            errors.append(
-                {
-                    "code": "region_unavailable",
-                    "message": "Expected citation region metadata was unavailable.",
-                    "expected": {
-                        "doc_id": doc_id,
-                        "page_number": page_number,
-                        "bbox": expected_bbox,
-                    },
-                }
-            )
-            continue
-        best_iou = max(bbox_iou(region["bbox"], expected_bbox) for region in candidate_regions)
-        if best_iou >= min_iou:
-            matched += 1
-            continue
-        errors.append(
-            {
-                "code": "region_misaligned",
-                "message": "Citation region bbox did not meet the IoU threshold.",
-                "expected": {
-                    "doc_id": doc_id,
-                    "page_number": page_number,
-                    "bbox": expected_bbox,
-                    "min_iou": min_iou,
-                },
-                "actual": {
-                    "best_iou": round(best_iou, 3),
-                    "regions": [
-                        {
-                            "page_number": region.get("page_number"),
-                            "bbox": region.get("bbox"),
-                            "block_id": region.get("block_id"),
-                        }
-                        for region in candidate_regions
-                    ],
-                },
-            }
-        )
-    return matched / len(expected_regions), errors
-
-
-def score_citation_grounding(
-    case: dict[str, Any],
-    prediction: dict[str, Any],
-) -> dict[str, Any]:
-    citations = answer_citations(prediction)
-    page_score, page_errors = score_citation_pages(case.get("expected_citation_pages") or [], citations)
-    region_score, region_errors = score_citation_regions(
-        case.get("expected_citation_regions") or [],
-        citations,
-    )
-    present_scores = [
-        score for score in (page_score, region_score) if isinstance(score, (int, float))
-    ]
-    return {
-        "citation_page_precision": page_score,
-        "citation_region_precision": region_score,
-        "citation_grounding": rate([float(score) for score in present_scores]),
-        "citation_grounding_errors": page_errors + region_errors,
-    }
-
-
-ALIGN_TOKEN_RE = re.compile(r"[A-Za-z0-9]+|[가-힣]+")
-ALIGN_STOPWORDS = {
-    "기관",
-    "핵심",
-    "요구사항",
-    "요구",
-    "차이",
-    "비교",
-    "관련",
-    "확인",
-    "대상",
-    "한다",
-    "있다",
-    "이다",
-    "은",
-    "는",
-    "이",
-    "가",
-    "을",
-    "를",
-    "의",
-    "와",
-    "과",
-}
-
-
-def normalized_alignment_text(text: str) -> str:
-    normalized = unicodedata.normalize("NFC", text).lower()
-    return re.sub(r"[^0-9a-z가-힣]+", "", normalized)
-
-
-def alignment_tokens(text: str) -> list[str]:
-    tokens = []
-    for match in ALIGN_TOKEN_RE.finditer(unicodedata.normalize("NFC", text).lower()):
-        token = match.group(0)
-        if len(token) <= 1 or token in ALIGN_STOPWORDS:
-            continue
-        tokens.append(token)
-    return tokens
-
-
-def evidence_alignment_text(item: dict[str, Any]) -> str:
-    parts = [
-        item.get("title", ""),
-        item.get("agency", ""),
-        item.get("project", ""),
-        item.get("section", ""),
-        item.get("text", ""),
-    ]
-    metadata = item.get("metadata")
-    if isinstance(metadata, dict):
-        for key, value in sorted(metadata.items()):
-            if value is None or value == "":
-                continue
-            parts.append(str(key))
-            parts.append(str(value))
-    return " ".join(str(part) for part in parts if str(part).strip())
-
-
-def claim_text_supported_by_citation(claim: dict[str, Any], citation_text: str) -> bool:
-    claim_text = " ".join(
-        str(claim.get(key) or "") for key in ("claim", "support") if claim.get(key)
-    )
-    if not claim_text.strip() or not citation_text.strip():
-        return False
-
-    compact_claim = normalized_alignment_text(str(claim.get("claim") or ""))
-    compact_citation = normalized_alignment_text(citation_text)
-    if compact_claim and compact_claim in compact_citation:
-        return True
-
-    tokens = alignment_tokens(str(claim.get("claim") or ""))
-    if not tokens:
-        return bool(compact_claim and compact_claim in compact_citation)
-    citation_tokens = set(alignment_tokens(citation_text))
-    overlap = sum(1 for token in tokens if token in citation_tokens)
-    return (overlap / max(1, len(tokens))) >= 0.5
-
-
-def expected_claim_specs_for_target(
-    case: dict[str, Any],
-    target: str,
-) -> list[dict[str, Any]]:
-    specs = case.get("expected_claim_citations") or []
-    matched = []
-    for spec in specs:
-        if not isinstance(spec, dict):
-            continue
-        expected_target = str(spec.get("target") or "").strip()
-        if not expected_target or expected_target == target:
-            matched.append(spec)
-    return matched
-
-
-def score_claim_citation_alignment(
-    case: dict[str, Any],
-    prediction: dict[str, Any],
-) -> dict[str, Any]:
-    evidence_by_chunk = {
-        str(item.get("chunk_id")): item
-        for item in prediction.get("evidence") or []
-        if item.get("chunk_id")
-    }
-    claims = answer_claims(prediction)
-    expected_specs = [
-        spec for spec in case.get("expected_claim_citations") or [] if isinstance(spec, dict)
-    ]
-    errors: list[dict[str, Any]] = []
-    scores: list[float] = []
-    claim_targets = {str(claim.get("target") or "") for claim in claims}
-
-    for claim_index, claim in enumerate(claims):
-        target = str(claim.get("target") or "")
-        citations = [citation for citation in claim.get("citations") or [] if isinstance(citation, dict)]
-        if not citations:
-            scores.append(0.0)
-            errors.append(
-                {
-                    "code": "claim_missing_citation",
-                    "message": "Claim has no citation.",
-                    "claim_index": claim_index,
-                    "target": target,
-                }
-            )
-            continue
-
-        citation_items = []
-        missing_chunk_ids = []
-        for citation in citations:
-            chunk_id = str(citation.get("chunk_id") or "")
-            item = evidence_by_chunk.get(chunk_id)
-            if item is None:
-                missing_chunk_ids.append(chunk_id)
-                continue
-            citation_items.append(item)
-        citation_text = " ".join(evidence_alignment_text(item) for item in citation_items)
-        citation_doc_ids = {str(item.get("doc_id") or "") for item in citation_items}
-
-        claim_errors: list[dict[str, Any]] = []
-        if missing_chunk_ids and not citation_items:
-            claim_errors.append(
-                {
-                    "code": "citation_not_in_evidence",
-                    "message": "Claim citation chunk was not present in top-level evidence.",
-                    "claim_index": claim_index,
-                    "target": target,
-                    "chunk_ids": missing_chunk_ids,
-                }
-            )
-        elif not claim_text_supported_by_citation(claim, citation_text):
-            claim_errors.append(
-                {
-                    "code": "claim_text_not_supported_by_citation",
-                    "message": "Claim text was not directly supported by the cited evidence text.",
-                    "claim_index": claim_index,
-                    "target": target,
-                    "citation_chunk_ids": [item.get("chunk_id") for item in citation_items],
-                }
-            )
-
-        for spec in expected_claim_specs_for_target(case, target):
-            expected_doc_ids = {str(doc_id) for doc_id in spec.get("expected_doc_ids") or []}
-            expected_terms = [str(term) for term in spec.get("expected_terms") or []]
-            if expected_doc_ids and not (expected_doc_ids & citation_doc_ids):
-                claim_errors.append(
-                    {
-                        "code": "expected_claim_doc_mismatch",
-                        "message": "Claim citation doc_id did not match the expected claim-level doc.",
-                        "claim_index": claim_index,
-                        "target": target,
-                        "expected_doc_ids": sorted(expected_doc_ids),
-                        "actual_doc_ids": sorted(doc_id for doc_id in citation_doc_ids if doc_id),
-                    }
-                )
-            if expected_terms and not contains_all_terms(citation_text, expected_terms):
-                claim_errors.append(
-                    {
-                        "code": "expected_claim_terms_missing",
-                        "message": "Claim citation text missed expected claim-level terms.",
-                        "claim_index": claim_index,
-                        "target": target,
-                        "expected_terms": expected_terms,
-                    }
-                )
-
-        errors.extend(claim_errors)
-        scores.append(0.0 if claim_errors else 1.0)
-
-    for spec in expected_specs:
-        expected_target = str(spec.get("target") or "").strip()
-        if expected_target and expected_target not in claim_targets:
-            scores.append(0.0)
-            errors.append(
-                {
-                    "code": "expected_claim_missing",
-                    "message": "Expected claim target was not emitted.",
-                    "target": expected_target,
-                }
-            )
-
-    return {
-        "claim_citation_alignment": rate(scores) if scores else None,
-        "claim_citation_aligned": sum(1 for score in scores if score >= 1.0),
-        "claim_citation_checked": len(scores),
-        "claim_citation_errors": errors,
-    }
-
-
-def score_answer_format(
-    case: dict[str, Any],
-    prediction: dict[str, Any],
-    policy: dict[str, Any] | None = None,
-) -> dict[str, Any]:
-    policy = policy or {}
-    answerable = bool(case.get("answerable", True))
-    expected_status = case.get("expected_answer_status")
-    if expected_status is None:
-        expected_status = (
-            policy.get("answerable_status", "supported")
-            if answerable
-            else policy.get("unanswerable_status", "insufficient")
-        )
-    min_claims = case.get("min_claims")
-    if min_claims is None:
-        min_claims = int(
-            policy.get("min_claims_answerable", 1)
-            if answerable
-            else policy.get("min_claims_unanswerable", 0)
-        )
-    require_claim_citations = bool(
-        case.get("require_claim_citations", policy.get("require_claim_citations", True))
-    )
-    expected_targets = {str(target) for target in case.get("expected_claim_targets") or []}
-    expected_missing_targets = {
-        str(target) for target in case.get("expected_missing_targets") or []
-    }
-
-    payload = answer_payload(prediction)
-    claims = answer_claims(prediction)
-    claim_targets = {str(claim.get("target") or "") for claim in claims}
-    citation_checks = []
-    for claim in claims:
-        citations = claim.get("citations") or []
-        citation_checks.append(
-            bool(citations)
-            and all(
-                isinstance(citation, dict)
-                and bool(citation.get("doc_id"))
-                and bool(citation.get("chunk_id"))
-                for citation in citations
-            )
-        )
-    citations_ok = True
-    if require_claim_citations and claims:
-        citations_ok = all(citation_checks)
-    elif require_claim_citations and int(min_claims) > 0:
-        citations_ok = False
-
-    insufficiency = payload.get("insufficiency") if isinstance(payload, dict) else {}
-    if not isinstance(insufficiency, dict):
-        insufficiency = {}
-    missing_targets = {str(target) for target in insufficiency.get("missing_targets") or []}
-
-    checks = {
-        "schema_version": int(payload.get("schema_version") or 0) >= 2,
-        "status_match": answer_status(prediction) == str(expected_status),
-        "min_claims": len(claims) >= int(min_claims),
-        "claim_targets": expected_targets.issubset(claim_targets),
-        "claim_citations": citations_ok,
-    }
-    if expected_missing_targets:
-        checks["missing_targets"] = expected_missing_targets.issubset(missing_targets)
-    return {
-        "expected_answer_status": str(expected_status),
-        "answer_status": answer_status(prediction),
-        "expected_claim_targets": sorted(expected_targets),
-        "claim_targets": sorted(target for target in claim_targets if target),
-        "expected_missing_targets": sorted(expected_missing_targets),
-        "missing_targets": sorted(target for target in missing_targets if target),
-        "claim_count": len(claims),
-        "format_checks": checks,
-        "answer_format_compliance": 1.0 if all(checks.values()) else 0.0,
-    }
-
-
-CHUNK_METRIC_KS = (5, 10)
-
-
-def derive_gold_chunk_ids(
-    case: dict[str, Any],
-    index: dict[str, Any] | None,
-) -> list[str]:
-    """Derive a chunk-level gold set from ``expected_doc_ids`` + ``expected_terms``.
-
-    A chunk is gold if its ``doc_id`` is in the case's ``expected_doc_ids`` AND
-    its text contains at least one ``expected_term``. If the case provides an
-    explicit ``gold_chunk_ids`` list it is used verbatim. Returns an empty list
-    for cases without expectations (e.g. abstention).
-    """
-    explicit = case.get("gold_chunk_ids")
-    if explicit:
-        return [str(item) for item in explicit if item]
-    expected_doc_ids = set(case.get("expected_doc_ids") or [])
-    expected_terms = [str(term) for term in case.get("expected_terms") or [] if term]
-    if not expected_doc_ids or not expected_terms or not index:
-        return []
-    chunks = index.get("chunks") or []
-    gold: list[str] = []
-    for chunk in chunks:
-        if chunk.get("doc_id") not in expected_doc_ids:
-            continue
-        text = str(chunk.get("text") or "")
-        if any(term in text for term in expected_terms):
-            chunk_id = str(chunk.get("chunk_id") or "")
-            if chunk_id:
-                gold.append(chunk_id)
-    return gold
-
-
-def chunk_recall_at_k(retrieved: list[str], gold: list[str], k: int) -> float | None:
-    if not gold:
-        return None
-    if not retrieved or k <= 0:
-        return 0.0
-    head = retrieved[:k]
-    hits = sum(1 for chunk_id in gold if chunk_id in head)
-    return hits / len(gold)
-
-
-def chunk_mrr(retrieved: list[str], gold: list[str]) -> float | None:
-    if not gold:
-        return None
-    if not retrieved:
-        return 0.0
-    gold_set = set(gold)
-    for rank, chunk_id in enumerate(retrieved, start=1):
-        if chunk_id in gold_set:
-            return 1.0 / rank
-    return 0.0
-
-
-def chunk_ndcg_at_k(retrieved: list[str], gold: list[str], k: int) -> float | None:
-    if not gold:
-        return None
-    if not retrieved or k <= 0:
-        return 0.0
-    gold_set = set(gold)
-    dcg = 0.0
-    for rank, chunk_id in enumerate(retrieved[:k], start=1):
-        rel = 1.0 if chunk_id in gold_set else 0.0
-        if rel:
-            dcg += rel / math.log2(rank + 1)
-    ideal_hits = min(len(gold_set), k)
-    idcg = sum(1.0 / math.log2(rank + 1) for rank in range(1, ideal_hits + 1))
-    return (dcg / idcg) if idcg else 0.0
-
-
-def score_case(
-    case: dict[str, Any],
-    prediction: dict[str, Any],
-    answer_policy: dict[str, Any] | None = None,
-    *,
-    gold_chunk_ids: list[str] | None = None,
-) -> dict[str, Any]:
-    answerable = bool(case.get("answerable", True))
-    query_type = canonical_query_type(case.get("query_type"))
-    expected_doc_ids = set(case.get("expected_doc_ids") or [])
-    expected_terms = [str(term) for term in case.get("expected_terms") or []]
-    expected_citation_terms = [
-        str(term) for term in case.get("expected_citation_terms") or expected_terms
-    ]
-    evidence = prediction.get("evidence") or []
-    evidence_doc_ids = {item.get("doc_id") for item in evidence}
-    answer = answer_to_text(prediction)
-    evidence_text = " ".join(str(item.get("text") or "") for item in evidence)
-    combined_text = " ".join([answer, evidence_text])
-    diagnostics = prediction.get("diagnostics") or {}
-    plan = prediction.get("plan") or {}
-    analysis = prediction.get("analysis") or {}
-    context_resolution = diagnostics.get("context_resolution") or {}
-    metadata_resolution = diagnostics.get("metadata_resolution") or {}
-    ambiguity = metadata_resolution.get("ambiguity") or {}
-    abstained = bool(diagnostics.get("abstained"))
-    answer_format = score_answer_format(case, prediction, answer_policy)
-    citation_grounding = score_citation_grounding(case, prediction)
-    claim_alignment = score_claim_citation_alignment(case, prediction)
-
-    citation_doc_precision = 0.0
-    if evidence_doc_ids:
-        citation_doc_precision = len(evidence_doc_ids & expected_doc_ids) / len(evidence_doc_ids)
-    citation_term_match = (
-        contains_all_terms(evidence_text, expected_citation_terms)
-        if expected_citation_terms
-        else bool(evidence)
-    )
-
-    comparison_target_recall: float | None = None
-    comparison_pool_recall: float | None = None
-    if query_type == "comparison" and len(expected_doc_ids) >= 2:
-        covered = expected_doc_ids & evidence_doc_ids
-        comparison_target_recall = len(covered) / len(expected_doc_ids)
-        coverage_after = (
-            ((prediction.get("plan") or {}).get("comparison_coverage") or {}).get("after") or {}
-        )
-        pool_doc_ids = {doc_id for doc_id, count in coverage_after.items() if count > 0}
-        if pool_doc_ids:
-            comparison_pool_recall = (
-                len(expected_doc_ids & pool_doc_ids) / len(expected_doc_ids)
-            )
-
-    if answerable:
-        doc_match = expected_doc_ids.issubset(evidence_doc_ids)
-        term_match = contains_all_terms(combined_text, expected_terms)
-        accuracy = 1.0 if doc_match and term_match and not abstained else 0.0
-        groundedness = 1.0 if term_match and evidence and not abstained else 0.0
-        citation_precision = citation_doc_precision if citation_term_match else 0.0
-        abstention = None
-    else:
-        doc_match = not evidence
-        term_match = abstained
-        accuracy = None
-        groundedness = 1.0 if abstained and not evidence else 0.0
-        citation_precision = 1.0 if abstained and not evidence else 0.0
-        abstention = 1.0 if abstained else 0.0
-
-    retrieved_chunk_ids = [
-        str(chunk_id)
-        for chunk_id in diagnostics.get("retrieved_chunk_ids") or []
-        if chunk_id
-    ]
-    gold_for_chunks = [str(item) for item in gold_chunk_ids or [] if item]
-    chunk_metrics: dict[str, float | None] = {
-        f"chunk_recall_at_{k}": chunk_recall_at_k(retrieved_chunk_ids, gold_for_chunks, k)
-        for k in CHUNK_METRIC_KS
-    }
-    chunk_metrics["chunk_mrr"] = chunk_mrr(retrieved_chunk_ids, gold_for_chunks)
-    chunk_metrics["chunk_ndcg_at_10"] = chunk_ndcg_at_k(retrieved_chunk_ids, gold_for_chunks, 10)
-
-    return {
-        "id": case.get("id"),
-        "query_type": query_type,
-        "slice": query_type,
-        "hardcase_categories": hardcase_categories(case),
-        "query": case.get("query"),
-        "answerable": answerable,
-        "expected_doc_ids": sorted(expected_doc_ids),
-        "evidence_doc_ids": sorted(doc_id for doc_id in evidence_doc_ids if doc_id),
-        "gold_chunk_ids": gold_for_chunks,
-        "retrieved_chunk_ids": retrieved_chunk_ids,
-        **chunk_metrics,
-        "doc_match": doc_match,
-        "term_match": term_match,
-        "citation_term_match": citation_term_match,
-        "citation_doc_precision": citation_doc_precision,
-        "accuracy": accuracy,
-        "groundedness": groundedness,
-        "citation_precision": citation_precision,
-        **citation_grounding,
-        **claim_alignment,
-        "abstention": abstention,
-        "comparison_target_recall": comparison_target_recall,
-        "comparison_pool_recall": comparison_pool_recall,
-        "latency_ms": diagnostics.get("latency_ms"),
-        "retry_count": diagnostics.get("retry_count", 0),
-        "retry_trigger_reasons": retry_trigger_reasons(prediction),
-        "last_attempt_verified": (
-            bool((diagnostics.get("filter_stage_attempts") or [])[-1].get("verified"))
-            if (diagnostics.get("filter_stage_attempts") or [])
-            else None
-        ),
-        "filter_stage": plan.get("filter_stage"),
-        "selected_top_k": diagnostics.get("selected_top_k"),
-        "retrieval_budget": dict(plan.get("retrieval_budget") or {}),
-        "metadata_ambiguous": bool(analysis.get("metadata_ambiguous")),
-        "ambiguity_decision": ambiguity.get("decision"),
-        "ambiguity_reason": ambiguity.get("decision_reason") or ambiguity.get("reason"),
-        "metadata_candidate_count": metadata_resolution.get("candidate_count"),
-        "metadata_selected_doc_ids": metadata_resolution.get("selected_doc_ids") or [],
-        "cold_start": bool(diagnostics.get("cold_start", False)),
-        "stage_latency": dict(diagnostics.get("stage_latency") or {}),
-        "attempt_latency": [
-            {
-                "stage": attempt.get("stage"),
-                "retrieve_ms": attempt.get("retrieve_ms", 0.0),
-                "verify_ms": attempt.get("verify_ms", 0.0),
-            }
-            for attempt in diagnostics.get("filter_stage_attempts") or []
-        ],
-        "context_resolution_status": context_resolution.get("status"),
-        "context_resolution_source": context_resolution.get("source"),
-        "context_resolution_confidence": context_resolution.get("confidence"),
-        "context_resolution_reason": context_resolution.get("reason"),
-        "resolved_query": prediction.get("resolved_query"),
-        "abstained": abstained,
-        **answer_format,
-        "answer": answer,
-        "evidence": [
-            {
-                "text": str(item.get("text") or "")[:600],
-                "doc_id": item.get("doc_id"),
-                "chunk_id": item.get("chunk_id"),
-                "page": item.get("page"),
-            }
-            for item in (prediction.get("evidence") or [])[:3]
-        ],
-    }
 
 
 _TOP_LEVEL_STAGE_KEYS = ("query_analysis_ms", "context_resolution_ms", "answer_generation_ms")
